# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Commutrum — JSE factor research engine API
servers:
  - url: /api/v1
    description: Versioned API base path
...
paths:
...
  /data/prices/{ticker}:
    get:
      operationId: getTickerPrices
      tags: [data]
      summary: Get daily OHLCV prices for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
        - $ref: "#/components/parameters/fromDate"
        - $ref: "#/components/parameters/toDate"
      responses:
        "200":
          description: Price series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceSeriesResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /data/fundamentals/{ticker}:
    get:
      operationId: getTickerFundamentals
      tags: [data]
      summary: Get point-in-time fundamentals for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
        - name: asOf
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Fundamentals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundamentalsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /data/events/{ticker}:
    get:
      operationId: getTickerEvents
      tags: [data]
      summary: Get SENS announcements and corporate events for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
        - $ref: "#/components/parameters/fromDate"
        - $ref: "#/components/parameters/toDate"
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsResponse"
  /data/quality/{ticker}:
    get:
      operationId: getTickerDataQuality
      tags: [data]
      summary: Get data quality summary for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
      responses:
        "200":
          description: Data quality summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataQualityResponse"

  # ── Module 02 — Factors ──────────────────────────────────────────────────
...
  /orthogonality/correlation-matrix:
    get:
      operationId: getCorrelationMatrix
      tags: [orthogonality]
      summary: Get pairwise factor correlation matrix
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Correlation matrix
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrelationMatrixResponse"
  /orthogonality/clusters:
    get:
      operationId: getSignalClusters
      tags: [orthogonality]
      summary: Get signal clusters and representatives
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Signal clusters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalClustersResponse"

  # ── Module 05 — Composite ────────────────────────────────────────────────
...
  /monitoring/alerts:
    get:
      operationId: getMonitoringAlerts
      tags: [monitoring]
      summary: Get open monitoring alerts
      parameters:
        - name: acknowledged
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertsResponse"
  /monitoring/alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeMonitoringAlert
      tags: [monitoring]
      summary: Acknowledge a monitoring alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcknowledgeRequest"
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcknowledgeResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Module 11 — Decay ────────────────────────────────────────────────────
  /decay/alerts:
    get:
      operationId: getDecayAlerts
      tags: [decay]
      summary: Get open decay alerts
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Decay alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecayAlertsResponse"
  /decay/alerts/{id}/investigate:
    post:
      operationId: investigateDecayAlert
      tags: [decay]
      summary: Log investigation findings for a decay alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestigationRequest"
      responses:
        "200":
          description: Investigation logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestigationResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Module 12 — Admission ────────────────────────────────────────────────
  /factor-candidates:
    get:
      operationId: getFactorCandidates
      tags: [admission]
      summary: List all factor candidates
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Factor candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactorCandidatesResponse"
  /factor-candidates/{id}/gate-results:
    get:
      operationId: getFactorCandidateGateResults
      tags: [admission]
      summary: Get all gate results for a factor candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gate results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateResultsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /factor-candidates/{id}/decision:
    post:
      operationId: recordFactorCandidateDecision
      tags: [admission]
      summary: Record admission decision for a factor candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionDecisionRequest"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionDecisionResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Module 13 — Registry ─────────────────────────────────────────────────
...
components:
  parameters:
    ticker:
      name: ticker
      in: path
      required: true
      schema:
        type: string
    fromDate:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date
    toDate:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Meta:
      type: object
      properties:
        as_of:
          type: string
          format: date
        generated_at:
          type: string
          format: date-time
      required: [generated_at]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    PricePoint:
      type: object
      properties:
        date:
          type: string
          format: date
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
        vwap:
          type: number
      required: [date, close]
    PriceSeriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    FundamentalsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            ticker:
              type: string
            periodEndDate:
              type: string
              format: date
            publicationDate:
              type: string
              format: date
            revenue:
              type: number
            netIncome:
              type: number
            totalAssets:
              type: number
            equity:
              type: number
            eps:
              type: number
            roe:
              type: number
            roa:
              type: number
          required: [ticker, periodEndDate, publicationDate]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    EventsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              category:
                type: string
              headline:
                type: string
            required: [date]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    DataQualityResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            ticker:
              type: string
            overallScore:
              type: number
            checks:
              type: array
              items:
                type: object
                properties:
                  checkType:
                    type: string
                  passed:
                    type: boolean
                  message:
                    type: string
                required: [checkType, passed]
          required: [ticker, overallScore]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...
    CorrelationMatrixResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              factorIdA:
                type: string
              factorIdB:
                type: string
              pearsonCorrelation:
                type: number
              vif:
                type: number
            required: [factorIdA, factorIdB, pearsonCorrelation]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    SignalClustersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              clusterId:
                type: string
              representative:
                type: string
              members:
                type: array
                items:
                  type: string
            required: [clusterId, representative, members]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...
    AlertEntry:
      type: object
      properties:
        id:
          type: integer
        alertDate:
          type: string
          format: date
        alertType:
          type: string
        factorId:
          type: string
        severity:
          type: string
        message:
          type: string
        acknowledged:
          type: boolean
      required: [id, alertDate, alertType, severity, message, acknowledged]
    AlertsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AlertEntry"
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    AcknowledgeRequest:
      type: object
      properties:
        acknowledgedBy:
          type: string
      required: [acknowledgedBy]
    AcknowledgeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: integer
            acknowledged:
              type: boolean
          required: [id, acknowledged]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    DecayAlertEntry:
      type: object
      properties:
        id:
          type: integer
        alertDate:
          type: string
          format: date
        factorId:
          type: string
        decayRule:
          type: string
        observedValue:
          type: number
        thresholdValue:
          type: number
        status:
          type: string
      required: [id, alertDate, factorId, decayRule, status]
    DecayAlertsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DecayAlertEntry"
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    InvestigationRequest:
      type: object
      properties:
        rootCauseCategory:
          type: string
          enum: [noise, structural_break, data_issue]
        findings:
          type: string
        decision:
          type: string
          enum: [retain, watch, remove]
        decisionRationale:
          type: string
      required: [rootCauseCategory, findings, decision, decisionRationale]
    InvestigationResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            alertId:
              type: integer
            decision:
              type: string
          required: [alertId, decision]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    FactorCandidateEntry:
      type: object
      properties:
        candidateId:
          type: string
        factorName:
          type: string
        causalityClassification:
          type: string
        status:
          type: string
        submittedAt:
          type: string
          format: date-time
      required: [candidateId, factorName, causalityClassification, status]
    FactorCandidatesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FactorCandidateEntry"
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    GateResultsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              gateNumber:
                type: integer
              gateName:
                type: string
              passed:
                type: boolean
              metricValue:
                type: number
              threshold:
                type: number
              evidence:
                type: string
            required: [gateNumber, gateName]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    AdmissionDecisionRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [admitted, rejected]
        rationale:
          type: string
        decidedBy:
          type: string
      required: [decision, rationale]
    AdmissionDecisionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            candidateId:
              type: string
            decision:
              type: string
          required: [candidateId, decision]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...